[refactor/#129] 이미지 다운로드 확장함수 적용 및 리스트 분리
